% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmv_proseq_hff_12hpi.R
\docType{data}
\name{cmv_proseq_hff_12hpi}
\alias{cmv_proseq_hff_12hpi}
\title{Sample PRO-seq dataset}
\format{
\subsection{\code{cmv_proseq_hff_12hpi}}{

A GRanges objects with 592079 fragments on the CMV genome (Genbank KF297339.1)
}
}
\source{
derived from NCBI's SRA data SRR16301856
}
\usage{
cmv_proseq_hff_12hpi
}
\description{
Sample BED file derived from a PRO-seq dataset measure nascent RNA transcribed from the CMV genome during an \emph{in vitro} CMV infection of fibroblast cells. PRO-seq was performed 12 hours post infection.
}
\keyword{datasets}
