% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmv_proseq_sample.R
\docType{data}
\name{cmv_proseq_sample}
\alias{cmv_proseq_sample}
\title{Sample PRO-seq dataset}
\format{
\subsection{\code{cmv_proseq_sample}}{

A GRanges objects with 7876 fragments on the CMV genome (Genbank KF297339.1)
}
}
\source{
derived from NCBI's SRA data SRR16301852
}
\usage{
cmv_proseq_sample
}
\description{
Sample BED file derived from a PRO-seq dataset measure nascent RNA transcribed from the CMV genome during an \emph{in vitro} CMV infection of fibroblast cells
}
\keyword{datasets}
