% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikein_normalization.R
\name{counts_in_genomicTiles}
\alias{counts_in_genomicTiles}
\title{Count coverage within genomic tiles}
\usage{
counts_in_genomicTiles(grlist, tilewidth = 50, average.method = "median")
}
\arguments{
\item{grlist}{a named GRangesList object representing the mapped reads for spike-in}

\item{tilewidth}{a positive integer giving the size of genomic tiles}

\item{average.method}{how to assess the average coverage within each genomic tile. Choice of "median" or "mean".  Or "none" if not averaging should be done.}
}
\value{
data.frame
}
\description{
Count coverage within genomic tiles
}
\examples{

n = 1000
chrLength = 1000
chrNames = c("chr1", "chr2")
mygenome = Seqinfo(
  seqnames = chrNames, 
  seqlengths = rep(chrLength, length(chrNames))) 
w = 10

gr1 <- GRanges(sample(chrNames,n, replace = TRUE), 
IRanges(sample(seq(chrLength-w), n, replace = TRUE), width = w), 
strand=sample(c("+", "-"), n, replace = TRUE),
seqinfo = mygenome)

gr2 <- GRanges(sample(chrNames,n, replace = TRUE), 
IRanges(sample(seq(chrLength-w), n, replace = TRUE), width = w), 
strand=sample(c("+", "-"), n, replace = TRUE),
seqinfo = mygenome)

grlist <- GRangesList(gr1, gr2)
names(grlist) <- c("sample_1", "sample_2")

counts_in_genomicTiles(grlist)
}
