% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tsr.R
\name{find_tsr}
\alias{find_tsr}
\title{Identify Transcription Start Regions}
\usage{
find_tsr(
  five_prime_grange,
  half_scoring_window_size,
  half_background_window_size,
  score.technique,
  pseudo_count = 1,
  threshold.technique = NULL,
  threshold.value = NULL
)
}
\arguments{
\item{five_prime_grange}{A GRanges object with all elements of width 1
represent the 5' location of PRO-seq reads}

\item{half_scoring_window_size}{The half window to score signal}

\item{half_background_window_size}{The window around each potential TSS to
assess the local background signal}

\item{score.technique}{A string identifying the technique to assess local
background signal}

\item{pseudo_count}{A pseudo-count when evaluating the ratio between signal
at a potential TSS and local-background signal}

\item{threshold.technique}{method for determining the threshold
score-to-background ratio that distinguishes a TSR from background}

\item{threshold.value}{Threshold value of signal-to-local background ratio to
call a TSS}
}
\value{
The five_prime_grange with two column added or replaced:
score_to_bg_ratio (a numeric) and TSS (a logical)
}
\description{
Identify Transcription Start Regions
}
\examples{
x = fragStartPosGRange(sample_NT2_data)
find_tsr(x, half_scoring_window_size = 50, half_background_window_size = 100,
score.technique = "mean", threshold.technique = "z-score", threshold.value = 1)
}
