% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tss.R
\name{find_tss}
\alias{find_tss}
\title{Identify Transcription Start Regions}
\usage{
find_tss(
  five_prime_grange,
  half_window_size,
  technique,
  pseudo_count = 1,
  threshold = NULL
)
}
\arguments{
\item{five_prime_grange}{A GRanges object with all elements of width 1 represent the 5' location of PRO-seq reads}

\item{half_window_size}{The window around each potential TSS to assess the local background signal}

\item{technique}{A string identifying the technique to assess local background signal}

\item{pseudo_count}{A pseudo-count when evaluating the ratio between signal at a potential TSS and local-background signal}

\item{threshold}{Threshold value of signal-to-local background ratio to call a TSS}
}
\value{
The five_prime_grange with two column added or replaced: score_to_bg_ratio (a numeric) and TSS (a logical)
}
\description{
Identify Transcription Start Regions
}
\examples{
x = make_fiveprimeends_grange(sample_NT2_data)
find_tss(x, half_window_size = 100, technique = "mean_exclusive")
}
