% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtsr_maxtss.R
\name{findtsr_maxtss}
\alias{findtsr_maxtss}
\title{Identify fixed windows that are transcription-start regions (TSRs) based on maximum value within each windows}
\usage{
findtsr_maxtss(x, w, background = 501, threshold = 0)
}
\arguments{
\item{x}{an Rle object}

\item{w}{width of window}

\item{background}{size of window to calculate background value. This should be significantly larger than w. To call a window as a TSR, the maximum value within a window must be greater than median of background window.}

\item{threshold}{the minimum value for maximum value within a window to call a TSR}
}
\value{
an IRanges object of all TSR windows
}
\description{
Identify fixed windows that are transcription-start regions (TSRs) based on maximum value within each windows
}
\examples{
x <- strand_coverage(cmv_proseq_sample)
x <- x[[1]][[1]]
findtsr_maxtss(x = x, w = 21, background = 501, threshold = 10)
}
