% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtsr_meantss.R
\name{findtsr_meantss}
\alias{findtsr_meantss}
\title{Identify fixed windows that are transcription-start regions (TSRs) based on sum value within each windows}
\usage{
findtsr_meantss(x, w, background = 501, threshold = 0)
}
\arguments{
\item{x}{an Rle object}

\item{w}{width of window}

\item{background}{size of window to calculate background value. This should be significantly larger than w. To call a window as a TSR, the mean value within a window must be greater than median value of background window.}

\item{threshold}{the minimum value for sum value of a window to call a TSR}
}
\value{
an IRanges object of all TSR windows
}
\description{
Identify fixed windows that are transcription-start regions (TSRs) based on sum value within each windows
}
\examples{
x <- strand_coverage(cmv_proseq_sample)
x <- x[[1]][[1]]
findtsr_meantss(x = x, w = 21, background = 501, threshold = 10)
}
