% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtsr_peaks.R
\name{findtsr_peaks}
\alias{findtsr_peaks}
\title{Find TSRs using a peaks based approach}
\usage{
findtsr_peaks(
  x,
  bg_size = 101,
  peak_pat = "[+]{1,}([0]{1,}[+]{1,}){0,}[-]{1,}([0]{1,}[-]{1,}){0,}",
  height_above_bg = 1,
  min_peak_distance = 1,
  min_gapwidth = 2
)
}
\arguments{
\item{x}{An S4Vectors::Rle object}

\item{bg_size}{The window around each position to assess background}

\item{peak_pat}{The regex pattern used to identify peaks. This is passed to pracma::findpeaks() function}

\item{height_above_bg}{The minimum (absolute) height a peak has to have to be recognized as such. This is passed to pracma::findpeaks() function}

\item{min_peak_distance}{the minimum distance (in indices) peaks have to have to be counted. This is passed to pracma::findpeaks() function}

\item{min_gapwidth}{the distance between two peaks to merge them into a single peak}
}
\value{
GRanges object
}
\description{
Find TSRs using a peaks based approach
}
\examples{
x <- Rle(c(rep(0,10),2,0,0,0,0,3,3,6,3,3,0,0,0,0,2,rep(0,10)))
findtsr_peaks(x, bg_size = 15, height_above_bg = 2)
}
