% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragStartPosGRange.R
\name{fragStartPosGRange}
\alias{fragStartPosGRange}
\title{Identify 5' end of all ranges in a GRanges object}
\usage{
fragStartPosGRange(input_bed)
}
\arguments{
\item{input_bed}{GRanges object}
}
\value{
GRanges object with all ranges of width 1
}
\description{
Identify 5' end of all ranges in a GRanges object
}
\examples{
fragStartPosGRange(sample_NT2_data)
}
