% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tss.R
\name{make_fiveprimeends_grange}
\alias{make_fiveprimeends_grange}
\title{Identify 5' end of all ranges in a GRanges object}
\usage{
make_fiveprimeends_grange(input_bed)
}
\arguments{
\item{input_bed}{GRanges object}
}
\value{
GRanges object with all ranges of width 1
}
\description{
Identify 5' end of all ranges in a GRanges object
}
\examples{
make_fiveprimeends_grange(sample_NT2_data)
}
