% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_coverage.R
\name{normalize_coverage}
\alias{normalize_coverage}
\title{Get normalized 5' end coverage for fragments mapped to genome}
\usage{
normalize_coverage(
  pos.cov,
  neg.cov,
  method = "total",
  normalization_factor = NULL
)
}
\arguments{
\item{pos.cov}{Rle object for positive strand}

\item{neg.cov}{Rle object for negative strand}

\item{method}{method by which to normalize. One of "total" (default) or
"numeric"}

\item{normalization_factor}{if method is "numeric", then this is the factor
by which coverage values are normalized}
}
\value{
a list of Rle objects. First item in list represents normalized
coverage of positive strand. Second item of list represents normalized
coverage of negative strand.
}
\description{
Get normalized 5' end coverage for fragments mapped to genome
}
\examples{
set.seed(100)

gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(
    start = sample(1:900, 100),
    width = sample(1:100, 100)),
  strand = sample(c("+", "-"), 100, replace = TRUE))

seqlengths(gr) <- 1000
x <- strand_coverage(gr)
normalize_coverage(x[[1]], x[[2]])
}
