% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_fiveprime_coverage.R
\name{normalized_fiveprime_coverage}
\alias{normalized_fiveprime_coverage}
\title{Get normalized 5' end coverage for fragments mapped to genome}
\usage{
normalized_fiveprime_coverage(gr)
}
\arguments{
\item{gr}{GRanges object with seqlength property included}
}
\value{
a list of Rle objects. First item in list represents normalized
coverage of positive strand. Second item of list represents normalized
coverage of negative strand.
}
\description{
at each position the normalized 5' coverage is calculated as
follows:
\deqn{x^2/y}
where \eqn{x} is the 5' coverage and \eqn{y} is the total fragment coverage
}
\examples{
set.seed(100)

gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(
    start = sample(1:900, 100),
    width = sample(1:100, 100)),
  strand = sample(c("+", "-"), 100, replace = TRUE))

seqlengths(gr) <- 1000
normalized_fiveprime_coverage(gr)
}
