% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfun.R
\name{runfun}
\alias{runfun}
\title{Apply a calculation through a sliding window across an rle object}
\usage{
runfun(x, width, FUN, ...)
}
\arguments{
\item{x}{Input Rle-object}

\item{width}{width of the sliding window}

\item{FUN}{the function to apply to the window}

\item{...}{any additional arguments to pass to zoo::rollapply}
}
\value{
Rle object same length as x
}
\description{
This function converts an Rle object to a vector, applies a sliding window function to it using zoo::rollapply.
}
\examples{
cmv_proseq_sample

x <- strand_coverage(cmv_proseq_sample)
runfun(x[[1]][[1]], width = 10, FUN = sum, fill = NA)
}
