% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_NT2_data}
\alias{sample_NT2_data}
\title{Sample PRO-seq Reads}
\format{
\subsection{\code{sample_NT2_data}}{

A GRanges object with 2184 reads and 2 metadata columns:
\describe{
\item{name}{name assigned to read by sequencer}
\item{score}{number of reads aligning to the range}
}
}
}
\source{
"In house experiment"
}
\usage{
sample_NT2_data
}
\description{
Random sample of PRO-seq reads aligning to CMV genome. Experiment was conducted
in D-NT2 cells infected with CMV
}
\keyword{datasets}
