% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikein_normalization.R
\name{spikein_normalization_factor}
\alias{spikein_normalization_factor}
\title{Get normalization factor for spikein controls}
\usage{
spikein_normalization_factor(grlist)
}
\arguments{
\item{grlist}{a named GRangesList object representing the mapped reads for spike-in}
}
\value{
data.frame
}
\description{
Get normalization factor for spikein controls
}
\examples{

n = 1000
chrLength = 1000
chrNames = c("chr1", "chr2")
mygenome = Seqinfo(
  seqnames = chrNames, 
  seqlengths = rep(chrLength, length(chrNames))) 
w = 10

gr1 <- GRanges(sample(chrNames,n, replace = TRUE), 
IRanges(sample(seq(chrLength-w), n, replace = TRUE), width = w), 
strand=sample(c("+", "-"), n, replace = TRUE),
seqinfo = mygenome)

gr2 <- GRanges(sample(chrNames,n, replace = TRUE), 
IRanges(sample(seq(chrLength-w), n, replace = TRUE), width = w), 
strand=sample(c("+", "-"), n, replace = TRUE),
seqinfo = mygenome)

grlist <- GRangesList(gr1, gr2)
names(grlist) <- c("sample_1", "sample_2")

spikein_normalization_factor(grlist)
}
