% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_coverage.R
\name{strand_coverage}
\alias{strand_coverage}
\title{Calculate strand specific coverage}
\usage{
strand_coverage(x)
}
\arguments{
\item{x}{Input GRanges object. Each interval must have either \code{+} or \code{-} assigned to its strand property}
}
\value{
A list of two Rle objects, the coverage on the positive strand and the coverage on the negative strand.
}
\description{
\code{strand_coverage} takes a GRanges object and returns a list of two Rle objects, the coverage on the positive strand and the coverage on the negative strand.
}
\examples{
# load a sample GRanges objects
cmv_proseq_sample

strand_coverage(cmv_proseq_sample)
}
