% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMetadata.R
\name{summarizeMetadata}
\alias{summarizeMetadata}
\title{Summarize GRanges metadata based on overlap with another GRange}
\usage{
summarizeMetadata(gr, grouping_gr, metadata, func = mean, ...)
}
\arguments{
\item{gr}{a granges object}

\item{grouping_gr}{a granges object}

\item{metadata}{charater indicating which metadata column in gr to summarize}

\item{func}{the function to use when summarizing metadata column. function must be able to}

\item{...}{additional arguments to pass to func}
}
\value{
grange with summarized metadata
}
\description{
Summarize GRanges metadata based on overlap with another GRange
}
\examples{

# Create dummy data
gr1 <- GRanges(
  seqnames = c(rep("chr1", 11)), strand = "+",
  ranges = IRanges(c(1000, 3000, 3600, 4000, 4000, 5000, 5400,
                     2000, 3000, 7000, 7500),
                   width = c(500, 500, 300, 500, 900, 500, 500,
                             900, 500, 600, 300))
)
gr2a <- resize(gr1, width = 50, fix = "start")
mcols(gr2a)$feature = sample(seq(100), length(gr2a))
gr2b <- resize(gr1, width = 50, fix = "center")
mcols(gr2b)$feature = sample(seq(100), length(gr2b))
gr2 <- c(gr2a, gr2b)

# Apply function
summarizeMetadata(gr2, gr1, metadata = "feature", func = mean)
}
