% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_grange.R
\name{bam_to_grange}
\alias{bam_to_grange}
\title{Title}
\usage{
bam_to_grange(x, keep)
}
\arguments{
\item{x}{GAlignmentPairs object}

\item{keep}{which metadata to keep}
}
\value{
GenomicRanges object
}
\description{
Title
}
\examples{
x <- system.file("inst", "extdata", "04_hpi_flavo_pro-seq_cmv_tb40e_sampled.bam")
bam_to_grange(x, keep = "mapq")
}
